+++ packages/site/examples/scene-case/default/demo/meta.json
@@ -91,14 +91,6 @@
         "en": "Why Do Cats?"
       },
       "screenshot": "https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*ug4vTJA7QbMAAAAAAAAAAAAADmJ7AQ/original"
-    },
-    {
-      "filename": "fishbone.js",
-      "title": {
-        "zh": "克服拖延症鱼骨图",
-        "en": "Anti-Procrastination Fishbone Diagram"
-      },
-      "screenshot": "https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*VusPQ50s3uEAAAAAAAAAAAAADmJ7AQ/original"
     }
   ]
 }
