+++ packages/site/examples/scene-case/tree-graph/demo/meta.json
@@ -1,0 +1,16 @@
+{
+  "title": {
+    "zh": "中文分类",
+    "en": "Category"
+  },
+  "demos": [
+    {
+      "filename": "anti-procrastination-fishbone.js",
+      "title": {
+        "zh": "克服拖延症鱼骨图",
+        "en": "Anti-Procrastination Fishbone Diagram"
+      },
+      "screenshot": "https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*VusPQ50s3uEAAAAAAAAAAAAADmJ7AQ/original"
+    }
+  ]
+}
