+++ packages/site/examples/scene-case/tree-graph/index.zh.md
@@ -1,0 +1,3 @@
+---
+title: 树图场景案例
+---
